feat: rename useUsernames
